🐛  fix : Main 페이지 컴포넌트 분할, 최상위 부모 상태를 관리로 인한 메인페이지 전체 리렌더링 방지
